Refactoring the widget interface.
Finished moving everything from http and https to request. Ideally the plugins
will only focus on fetching data, not worrying about sending it to clients.
Now the last updates are cached and sent immediately on connection.

I need to figure out the best way for the widgets to tell the app how frequently they
should be updated. Also need to figure out a way to have widgets hook themselves
in to broadcast so I can get the tweet streaming working again.

Also it might make more sense to compute the phase of the moon on the server
since it doesn't change often.